Create tools files for organization
