+++ Block Based TalkThru (C)/Block Based TalkThru.dpj
@@ -71,6 +71,8 @@
 			<files>
 				<file name=".\ad1835.h">
 				</file>
+				<file name=".\tools.h">
+				</file>
 				<file name=".\tt.h">
 				</file>
 			</files>
@@ -192,6 +194,22 @@
 					</file-configurations>
 				</file>
 				<file name=".\SPORTisr.c">
+					<file-configurations>
+						<file-configuration name="Debug">
+							<excluded-flag value="no"/>
+							<build-with-flag value="project"/>
+							<intermediate-dir>.\Debug</intermediate-dir>
+							<output-dir>.\Debug</output-dir>
+						</file-configuration>
+						<file-configuration name="Release">
+							<excluded-flag value="no"/>
+							<build-with-flag value="project"/>
+							<intermediate-dir>.\Release</intermediate-dir>
+							<output-dir>.\Release</output-dir>
+						</file-configuration>
+					</file-configurations>
+				</file>
+				<file name=".\tools.c">
 					<file-configurations>
 						<file-configuration name="Debug">
 							<excluded-flag value="no"/>
