+++ Block Based TalkThru (C)/tools.h
@@ -1,0 +1,11 @@
+#ifndef TOOLS_H
+#define TOOLS_H
+
+// returns the absolute value of a float
+float absolute(float val);
+
+// returns the approximation of exp(x)
+// based on the first n temrs of Taylor series
+float exponential(float x, int n);
+
+#endif // TOOLS_H
